+++ NEWS.md
@@ -2,6 +2,10 @@
 
 This is a major release that contains a significant number of breaking changes. For users who have been using the package for a while we recommend reading the [migration guide](https://epiforecasts.io/EpiNow2/articles/migration.html) to help with the transition.
 
+## Features
+
+* Generalises `estimate_truncation()` to work on data of any frequency (e.g. weekly). By @sbfnk in XXX.
+
 # EpiNow2 1.3.4
 
 This release focusses on bug fixes and package infrastructure updates along with a few quality of life improvements such as enabling the use of fixed delays and generation times.